Move `CodegenBackendKind` to its own module `crate::core::backend`
