+++ src/bootstrap/src/core/config/flags.rs
@@ -10,13 +10,14 @@ use clap_complete::Generator;
 #[cfg(feature = "tracing")]
 use tracing::instrument;
 
+use crate::core::backend::CodegenBackendKind;
 use crate::core::build_steps::perf::PerfArgs;
 use crate::core::build_steps::setup::Profile;
 use crate::core::builder::{Builder, Kind};
 use crate::core::config::Config;
 use crate::core::config::target_selection::{TargetSelectionList, target_selection_list};
 use crate::utils::helpers;
-use crate::{Build, CodegenBackendKind, TestTarget};
+use crate::{Build, TestTarget};
 
 #[derive(Copy, Clone, Default, Debug, ValueEnum)]
 pub enum Color {
