+++ src/bootstrap/src/core/config/toml/rust.rs
@@ -7,7 +7,7 @@ use std::path::PathBuf;
 use build_helper::ci::CiEnv;
 use serde::{Deserialize, Deserializer};
 
-use crate::CodegenBackendKind;
+use crate::core::backend::CodegenBackendKind;
 use crate::core::config::macros::define_config;
 use crate::core::config::toml::TomlConfig;
 use crate::core::config::{CompressDebuginfo, DebuginfoLevel, StringOrBool, TargetSelection};
@@ -430,6 +430,7 @@ pub fn check_incompatible_options_for_ci_rustc(
 
 pub(crate) const BUILTIN_CODEGEN_BACKENDS: &[&str] = &["llvm", "cranelift", "gcc"];
 
+/// FIXME(Zalathar): This is partly redundant with the parsing code in [`CodegenBackendKind`].
 pub(crate) fn parse_codegen_backends(
     backends: Vec<String>,
     section: &str,
