+++ src/bootstrap/src/core/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod android;
+pub(crate) mod backend;
 pub(crate) mod build_steps;
 pub(crate) mod builder;
 pub(crate) mod compiler;
