+++ src/bootstrap/src/lib.rs
@@ -51,60 +51,6 @@ pub mod cli_main;
 mod core;
 mod utils;
 
-/// Represents a codegen backend.
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Default)]
-pub enum CodegenBackendKind {
-    #[default]
-    Llvm,
-    Cranelift,
-    Gcc,
-    Custom(String),
-}
-
-impl CodegenBackendKind {
-    /// Name of the codegen backend, as identified in the `compiler` directory
-    /// (`rustc_codegen_<name>`).
-    pub fn name(&self) -> &str {
-        match self {
-            CodegenBackendKind::Llvm => "llvm",
-            CodegenBackendKind::Cranelift => "cranelift",
-            CodegenBackendKind::Gcc => "gcc",
-            CodegenBackendKind::Custom(name) => name,
-        }
-    }
-
-    /// Name of the codegen backend's crate, e.g. `rustc_codegen_cranelift`.
-    pub fn crate_name(&self) -> String {
-        format!("rustc_codegen_{}", self.name())
-    }
-
-    pub fn is_llvm(&self) -> bool {
-        matches!(self, Self::Llvm)
-    }
-
-    pub fn is_cranelift(&self) -> bool {
-        matches!(self, Self::Cranelift)
-    }
-
-    pub fn is_gcc(&self) -> bool {
-        matches!(self, Self::Gcc)
-    }
-}
-
-impl std::str::FromStr for CodegenBackendKind {
-    type Err = &'static str;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s.to_lowercase().as_str() {
-            "" => Err("Invalid empty backend name"),
-            "gcc" => Ok(Self::Gcc),
-            "llvm" => Ok(Self::Llvm),
-            "cranelift" => Ok(Self::Cranelift),
-            _ => Ok(Self::Custom(s.to_string())),
-        }
-    }
-}
-
 #[derive(PartialEq, Eq, Copy, Clone, Debug)]
 pub enum TestTarget {
     /// Run unit, integration and doc tests (default).
